+++ LittleCompany/events/enemy/EnemyEventManager.cs
@@ -1,14 +1,11 @@
 ﻿using System.Collections.Generic;
 using System;
-using UnityEngine;
-
-using LittleCompany.modifications;
-using LittleCompany.helper;
-using static LittleCompany.helper.EnemyInfo;
-using Unity.Netcode;
 using System.Collections;
-using GameNetcodeStuff;
+using UnityEngine;
+using LittleCompany.helper;
 using LittleCompany.components;
+using static LittleCompany.helper.EnemyInfo;
+using LittleCompany.events.item;
 
 namespace LittleCompany.events.enemy
 {
@@ -45,9 +42,10 @@ namespace LittleCompany.events.enemy
             return null;
         }
 
-        public static void BindAllEnemyEvents()
+        public static void BindAllEventHandler()
         {
             int handlersAdded = 0;
+            int customHandlersAdded = 0;
             foreach (var enemyType in EnemyTypes)
             {
                 if (enemyType.enemyPrefab == null)
@@ -64,12 +62,20 @@ namespace LittleCompany.events.enemy
                         handlersAdded++;
 #if DEBUG
                     if (eventHandler != null)
-                        Plugin.Log("Added event handler \"" + eventHandlerType.Name + "\" for enemy \"" + enemyType.enemyName + "\"");
+                    {
+                        if (eventHandlerType == typeof(CustomEnemyEventHandler))
+                            customHandlersAdded++;
+                        else
+                            Plugin.Log("Added event handler \"" + eventHandlerType.Name + "\" for enemy \"" + enemyType.enemyName + "\"");
+                    }
                     else
                         Plugin.Log("No enemy handler found for enemy \"" + enemyType.enemyName + "\"");
 #endif
                 }
             }
+
+            if (customHandlersAdded > 0)
+                Plugin.Log("Added custom event handler for " + customHandlersAdded + " enemies");
 
             Plugin.Log("BindAllEnemyEvents -> Added handler for " + handlersAdded + "/" + EnemyTypes.Count + " enemies.");
         }
@@ -80,17 +86,14 @@ namespace LittleCompany.events.enemy
             RobotEventHandler.LoadBurningRobotToyPrefab();
         }
 
-        public class EnemyEventHandler : NetworkBehaviour, IScalingListener
+        public class EnemyEventHandler : EventHandlerBase
         {
             internal EnemyAI enemy = null;
-            internal float DeathPoofScale = Effects.DefaultDeathPoofScale;
 
-            void Awake()
+            public override void OnAwake()
             {
-                Plugin.Log(name + " event handler has awaken!");
                 enemy = GetComponent<EnemyAI>();
-
-                OnAwake();
+                GetComponent<EnemyScaling>()?.AddListener(this);
 #if DEBUG
                 //StartCoroutine(SpawnKillLater());
 #endif
@@ -105,78 +108,16 @@ namespace LittleCompany.events.enemy
             }
 #endif
 
-            public virtual void OnAwake() {
-                GetComponent<EnemyScaling>()?.AddListener(this);
-            }
-
-            public void AtEndOfScaling(float from, float to, PlayerControllerB playerBy)
+            public override void DestroyObject()
             {
-                if (Mathf.Approximately(from, to)) return;
-                Scaled(from, to, playerBy);
-				if (from > to)
-                    Shrunken(from <= 1f, playerBy);
-                else
-                    Enlarged(from >= 1f, playerBy);
-            }
-
-            public virtual void AboutToDeathShrink(float currentSize, PlayerControllerB playerShrunkenBy) {
-                Plugin.Log("AboutToDeathShrink");
-            }
-
-            public virtual void OnDeathShrinking(float previousSize, PlayerControllerB playerShrunkenBy)
-            {
-
-                if (Effects.TryCreateDeathPoofAt(out _, enemy.transform.position, DeathPoofScale) && enemy.gameObject.TryGetComponent(out AudioSource audioSource) && audioSource != null && Modification.deathPoofSFX != null)
-                    audioSource.PlayOneShot(Modification.deathPoofSFX);
-
-                if (PlayerInfo.IsHost)
-                    StartCoroutine(DespawnAfterEventSync());
-                else
-                    DeathShrinkEventReceivedServerRpc();
-
-                Plugin.Log("Enemy shrunken to death");
-            }
-
-            #region DeathShrinkSync
-            // Has to be done to prevent the enemy from despawning before any client got the OnDeathShrinking event
-            private int DeathShrinkSyncedPlayers = 1; // host always got it
-            public IEnumerator DespawnAfterEventSync()
-            {
-                var waitedFrames = 0;
-                var playerCount = PlayerInfo.AllPlayers.Count;
-                Plugin.Log("EnemyEventManager.PlayerCount: " + playerCount);
-                while (waitedFrames < 100 && DeathShrinkSyncedPlayers < playerCount)
-                {
-                    waitedFrames++;
-                    yield return null;
-                }
-
-                if (waitedFrames == 100)
-                    Plugin.Log("Timeout triggered the death shrink event.", Plugin.LogType.Warning);
-                else
-                    Plugin.Log("Syncing the death shrink event took " + waitedFrames + " frames.");
-
-                DeathShrinkSyncedPlayers = 1;
-
-                // Now we can despawn it
                 enemy.KillEnemyServerRpc(true);
                 enemy.enabled = false;
-                yield return new WaitForSeconds(3f);
+            }
+
+            public override void DespawnObject()
+            {
                 RoundManager.Instance.DespawnEnemyOnServer(enemy.NetworkObject);
             }
-
-            [ServerRpc(RequireOwnership = false)]
-            public void DeathShrinkEventReceivedServerRpc()
-            {
-                DeathShrinkSyncedPlayers++;
-            }
-            #endregion
-
-			public virtual void Scaled(float from, float to, PlayerControllerB playerShrunkenBy) { }
-
-            public virtual void Shrunken(bool wasAlreadyShrunken, PlayerControllerB playerShrunkenBy) { }
-            public virtual void Enlarged(bool wasAlreadyEnlarged, PlayerControllerB playerEnlargedBy) { }
-            public virtual void ScaledToNormalSize(bool wasShrunken, bool wasEnlarged, PlayerControllerB playerScaledBy) { }
         }
     }
 }