+++ LittleCompany/events/item/CustomItemEventHandler.cs
@@ -1,0 +1,8 @@
+﻿using static LittleCompany.events.item.ItemEventManager;
+
+namespace LittleCompany.events.item
+{
+    internal class CustomItemEventHandler : ItemEventHandler
+    {
+    }
+}